代码格式优化

+++ src/main/java/org/kin/framework/utils/FileUtils.java
@@ -1,6 +1,7 @@
 package org.kin.framework.utils;
 
 import java.io.File;
+import java.util.Objects;
 
 /**
  * @author huangjianqin
@@ -13,13 +14,16 @@ public class FileUtils {
     }
 
     public static boolean delete(File targetFile) {
-        if (targetFile.isDirectory()) {
-            for (File childFile : targetFile.listFiles()) {
-                delete(childFile);
+        if (Objects.nonNull(targetFile)) {
+            if (targetFile.isDirectory()) {
+                for (File childFile : targetFile.listFiles()) {
+                    delete(childFile);
+                }
+                return targetFile.delete();
+            } else {
+                return targetFile.delete();
             }
-            return targetFile.delete();
-        } else {
-            return targetFile.delete();
         }
+        return false;
     }
 }
